tests/Elksoft.Converters.Tests/Elksoft.Converters.BooleanConverters/BooleanToUInt16ConverterTests.cs
using System;
using FluentAssertions;
using Moq;
using Xunit;

namespace Elksoft.Converters.BooleanConverters
{
    public class BooleanToUInt16ConverterTests
    {
        [Fact]
        public static void AcceptsNull_IsFalse()
        {
            var converter = new BooleanToUInt16Converter();
            converter.AcceptsNull.Should().BeFalse();
        }

        [Fact]
        public static void IsExplicit_IsFalse()
        {
            var converter = new BooleanToUInt16Converter();
            converter.IsExplicit.Should().BeFalse();
        }

        public static TheoryData<Boolean, UInt16> Convert_ReturnsExpected_Data()
        {
            return new TheoryData<Boolean, UInt16>() {
                { false, (UInt16)0 },
                { true, (UInt16)1 },
            };
        }

        [Theory]
        [MemberData(nameof(Convert_ReturnsExpected_Data))]
        public static void Convert_ReturnsExpected(Boolean input, UInt16 output)
        {
            var mockProvider = new Mock<IFormatProvider>(MockBehavior.Strict);

            var converter = new BooleanToUInt16Converter();
            converter.Convert(input, null).Should().Be(output);
            converter.Convert(input, mockProvider.Object).Should().Be(output);

            mockProvider.VerifyAll();
        }
    }
}
